requirements.txt
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
numpy>=1.24.0
tabulate>=0.9.0
